update return types

+++ shelkovnikov.dmitrii/I2/countmaximumelements.h
@@ -1,5 +1,5 @@
 #ifndef SPBSPU_LABS_2022_AAP_904_A_COUNTMAXIMUMELEMENTS_H
 #define SPBSPU_LABS_2022_AAP_904_A_COUNTMAXIMUMELEMENTS_H
 #include <cstddef>
-unsigned int count_maximum_elements(const int *arr, const size_t size);
+size_t count_maximum_elements(const int *arr, const size_t size);
 #endif
