+++ shelkovnikov.dmitrii/I2/countorderedelements.h
@@ -1,5 +1,5 @@
 #ifndef SPBSPU_LABS_2022_AAP_904_A_COUNTORDEREDELEMENTS_H
 #define SPBSPU_LABS_2022_AAP_904_A_COUNTORDEREDELEMENTS_H
 #include <cstddef>
-unsigned int count_ordered_elements(const int *arr, const size_t size);
+size_t count_ordered_elements(const int *arr, const size_t size);
 #endif
